Python backdoor
This is a very simple and basic python backdoor program which is intended to help those who are new to python programing and are aspiring to be white hat hackers like myself. Am intending to improve the program with time, This is my fast project and am going to make it one of the best backdoors ever for ethical hacking so I welcome any suggestions for modification and possible changes within it as aI improve it.